requirements.txt
Flask
Pillow
numpy
gunicorn
flask-cors
